pyproject.toml
[project]
name = "amen-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "aiosmtplib>=4.0.2",
    "alembic>=1.17.2",
    "fastapi>=0.116.2",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.11.9",
    "pyjwt[crypto]>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "redis[hiredis]>=6.4.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.35.0",
]
